skip grumpyscreen if helixscreen init script is detected

+++ k1/switch-to-stock.sh
@@ -66,8 +66,16 @@ if [ -f /usr/data/backups/creality-backup.tar.gz ]; then
       rm -rf /usr/data/printer_data/config/*.cfg
       rm -rf /usr/data/printer_data/config/*.conf
 
-      # for stock screen grumpyscreen needs to be disabled
-      rm /etc/init.d/S99guppyscreen
+      # for stock screen helixscreen/grumpyscreen needs to be disabled
+      if [ -f /etc/init.d/S99helixscreen ]; then
+        echo
+        echo "INFO: HelixScreen has been disabled"
+        mv /etc/init.d/S99helixscreen /usr/data
+      elif [ -f /etc/init.d/S99guppyscreen ]; then
+        echo
+        echo "INFO: GrumpyScreen has been disabled"
+        rm /etc/init.d/S99guppyscreen
+      fi
 
       # need these files restored back so that moonraker starts correctly
       cp /usr/data/pellcorp/k1/moonraker.conf /usr/data/printer_data/config/
@@ -126,7 +134,18 @@ EOF
       rm -rf /usr/data/printer_data/config/*.cfg
       rm -rf /usr/data/printer_data/config/*.conf
       cp /usr/data/pellcorp/k1/services/S55klipper_service /etc/init.d/
-      cp /usr/data/pellcorp/k1/services/S99guppyscreen /etc/init.d/
+
+      # the backed up /etc/init.d/S99helixscreen from switching to stock
+      if [ -f /usr/data/S99helixscreen ]; then
+        echo
+        echo "INFO: Restoring HelixScreen"
+        mv /usr/data/S99helixscreen /etc/init.d/
+      else
+        echo
+        echo "INFO: Restoring GrumpyScreen"
+        cp /usr/data/pellcorp/k1/services/S99guppyscreen /etc/init.d/
+      fi
+
       ln -sf /usr/data/klipper/fw/K1/klipper_host_mcu /usr/bin/klipper_mcu
       /usr/data/pellcorp/tools/backups.sh --restore backup-latest.tar.gz
       [ -f /root/.profile.stock ] && rm /root/.profile.stock
